% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurrecon.R
\name{hurrecon_model_site}
\alias{hurrecon_model_site}
\alias{hurrecon_model_site_all}
\alias{hurrecon_model_region}
\alias{hurrecon_model_region_all}
\title{Modeling Functions}
\usage{
hurrecon_model_site(
  hur_id,
  site_name,
  width = FALSE,
  time_step = 1,
  save = TRUE,
  timing = TRUE
)

hurrecon_model_site_all(
  site_name,
  width = FALSE,
  time_step = 1,
  save = TRUE,
  timing = TRUE
)

hurrecon_model_region(
  hur_id,
  width = FALSE,
  time_step = NULL,
  water = FALSE,
  save = TRUE,
  timing = TRUE
)

hurrecon_model_region_all(width = FALSE, time_step = NULL, water = FALSE)
}
\arguments{
\item{hur_id}{hurricane id}

\item{site_name}{name of site}

\item{width}{whether to use width parameters for the specified hurricane}

\item{time_step}{time step (minutes)}

\item{save}{whether to save results to a GeoTiff file}

\item{timing}{whether to display total elapsed time}

\item{water}{whether to calculate results over water}
}
\value{
a data frame of results if save is FALSE

a data frame of results if save is FALSE

a brick of 6 rasters if save is FALSE

no return value
}
\description{
hurrecon_model_site calculates wind speed (meters/second), gust speed 
(meters/second), wind direction (degrees), and enhanced Fujita scale wind 
damage for a given hurricane and site. If width is TRUE, the radius of 
maximum wind (rmw) and profile exponent (s_par) for the given hurricane are
used, if available. If save is TRUE, results are saved to a CSV file on 
the site subdirectory; otherwise results are returned as a data frame.  
If timing is TRUE, the total elasped time is displayed.

hurrecon_model_site_all creates a table of peak values for all hurricanes
for a given site. If width is TRUE, the radius of maximum wind (rmw) and 
profile exponent (s_par) for the given hurricane are used, if available. 
If save is TRUE, results are saved to a CSV file on the site subdirectory; 
otherwise results are returned as a data frame.  If timing is TRUE, the 
total elasped time is displayed.

hurrecon_model_region calculates peak wind speed (meters/second), enhanced
Fujita scale, wind direction (degrees), cardinal wind direction, gale wind
duration (minutes), and hurricane wind duration (minutes) for a given 
hurricane over a region. If width is TRUE, the radius of maximum wind (rmw) 
and profile exponent (s_par) for the given hurricane are used, if available.
If time_step is NULL, the time step is calculated. If water is FALSE, results
are calculated for land areas only. If save is TRUE, results are saved as a 
GeoTiff file; otherwise results are returned as a raster brick with 6 layers.
If timing is TRUE, the total elasped time is displayed.

hurrecon_model_region_all calculates peak wind speed (meters/second), 
enhanced Fujita scale, wind direction (degrees), cardinal wind direction, 
duration of gale winds (minutes), and duration of hurricane winds (minutes) 
over a region for all hurricanes. If width is TRUE, the radius of maximum 
wind (rmw) and profile exponent (s_par) for the given hurricane are used, 
if available. If time_step is NULL, the time step is calculated. If water 
is FALSE, results are calculated for land areas only. Results for each 
hurricane are saved in a GeoTiff file on the region subdirectory. Summary
results for all hurricanes (summary.tif, summary.csv) are also calculated 
and saved to the region subdirectory.
}
