% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurrecon.R
\name{hurrecon_plot_site_ts}
\alias{hurrecon_plot_site_ts}
\alias{hurrecon_plot_site_xy}
\alias{hurrecon_plot_site_all}
\alias{hurrecon_plot_region}
\alias{hurrecon_plot_region_all}
\title{Plotting Functions}
\usage{
hurrecon_plot_site_ts(
  hur_id,
  site_name,
  start_datetime = "",
  end_datetime = "",
  y_var = "wind_spd"
)

hurrecon_plot_site_xy(
  hur_id,
  site_name,
  start_datetime = "",
  end_datetime = "",
  y_var = "wind_spd",
  adjust = FALSE
)

hurrecon_plot_site_all(
  site_name,
  start_year = "",
  end_year = "",
  y_var = "wind_spd"
)

hurrecon_plot_region(hur_id, var = "fujita_scale")

hurrecon_plot_region_all(var = "efmax", tracks = FALSE)
}
\arguments{
\item{hur_id}{hurricane id}

\item{site_name}{name of site}

\item{start_datetime}{optional start datetime in format YYYY-MM-DD hh:mm}

\item{end_datetime}{optional end datetime in format YYYY-MM-DD hh:mm}

\item{y_var}{y-axis variable}

\item{adjust}{whether to subtract 360 degrees from wind direction.}

\item{start_year}{optional start year}

\item{end_year}{optional end year}

\item{var}{variable to plot}

\item{tracks}{whether to also plot hurricane tracks}
}
\value{
no return value

no return value

no return value

no return value

no return value
}
\description{
hurrecon_plot_site_ts creates a time-series plot of wind speed, gust 
speed, or wind direction as a function of datetime for a given 
hurricane and site. Optional start and end datetimes may be specified.

hurrecon_plot_site_xy creates a scatter plot of wind speed or gust speed
as a function of wind direction for a given hurricane and site. If adjust 
is TRUE, 360 degrees are substracted from wind direction values greater 
than 180. Optional start and end datetimes may be specified.

hurrecon_plot_site_all creates a time-series plot of peak values 
for all hurricanes for a given site. Optional start and end years
may be specified.

hurrecon_plot_region creates regional plots of enhanced Fujita scale, 
peak wind speed, wind direction, cardinal wind direction, gale wind 
duration, and hurricane wind duration for a given hurricane.

hurrecon_plot_region_all creates regional plots of maximum enhanced Fujita
value and number of storms for each enhanced Fujita value for all hurricanes.
}
