% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurrecon.R
\name{hurrecon_summarize_land_water}
\alias{hurrecon_summarize_land_water}
\alias{hurrecon_summarize_tracks}
\alias{hurrecon_summarize_site}
\title{Summarizing Functions}
\usage{
hurrecon_summarize_land_water()

hurrecon_summarize_tracks()

hurrecon_summarize_site(hur_id, site_name)
}
\arguments{
\item{hur_id}{hurricane id}

\item{site_name}{name of site}
}
\value{
no return value

no return value

no return value
}
\description{
hurrecon_summarize_land_water displays features of the current land-water
file (land_water.tif) in the console.

hurrecon_summarize_tracks displays features of the current ids file 
(ids.csv) in the console.

hurrecon_summarize_site displays peak values for a given hurricane
and site in the console.
}
